Moved Android-related heavy lifting from Gradle plugin to separate module
