+++ meta/gradle-plugin/build.gradle.kts
@@ -16,9 +16,6 @@ dependencies {
         exclude("org.jetbrains.kotlin", "kotlin-native-library-reader")
     }
 
-    implementation("androidx.annotation:annotation:1.1.0")
-    implementation("org.ow2.asm:asm:7.2")
-    implementation("org.ow2.asm:asm-tree:7.2")
     implementation("com.jfrog.bintray.gradle:gradle-bintray-plugin:1.8.4")
     compileClasspath("com.android.tools.build:gradle:3.5.2")
     runtimeClasspath("com.android.tools.build:gradle:3.5.2")
