--- meta/gradle-plugin/src/main/kotlin/constants.kt
@@ -1,5 +1,0 @@
-package dev.inkremental.meta.gradle
-
-internal const val PACKAGE = "trikita.anvil"
-internal const val VIEW_CNAME = "android.view.View"
-internal const val ROOT_VIEW_SCOPE = "RootViewScope"
