+++ meta/model/src/main/kotlin/Introspector.kt
@@ -1,0 +1,5 @@
+package dev.inkremental.meta.model
+
+interface Introspector {
+    fun viewModelSequence(quirks: InkrementalQuirks): Sequence<Pair<ViewModel, List<String>>>
+}
