+++ meta/model/src/main/kotlin/types.kt
@@ -1,0 +1,7 @@
+package dev.inkremental.meta.model
+
+// Inner Any can be String and Boolean for __viewAlias and Boolean for others
+typealias InkrementalQuirks = MutableMap<String, Map<String, Any?>>
+
+enum class InkrementalType { SDK, LIBRARY }
+enum class InkrementalPlatform { ANDROID, IOS }
