+++ meta/model/src/main/kotlin/utils.kt
@@ -1,0 +1,5 @@
+package dev.inkremental.meta.model
+
+import java.io.File
+
+operator fun File.div(relative: String): File = resolve(relative)
